move test fixture for live tests to live tests folder
