+++ packages/react-composites/tests/browser/chat/live-tests/fixture.ts
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { TEST_PARTICIPANTS_CHAT } from '../common/constants';
-import { createChatUsers, usePagePerParticipant } from '../common/fixtureHelpers';
-import { WorkerFixture, ChatUserType } from '../common/fixtureTypes';
-import { createTestServer } from '../../server';
+import { TEST_PARTICIPANTS_CHAT } from '../../common/constants';
+import { createChatUsers, usePagePerParticipant } from '../../common/fixtureHelpers';
+import { WorkerFixture, ChatUserType } from '../../common/fixtureTypes';
+import { createTestServer } from '../../../server';
 import { test as base } from '@playwright/test';
 import path from 'path';
